+++ src/services/provider/geoResource.provider.js
@@ -3,7 +3,7 @@
  */
 import { AggregateGeoResource, VectorGeoResource, WmsGeoResource, WMTSGeoResource, VectorSourceType } from '../domain/geoResources';
 import { $injector } from '../../injection';
-import { getBvvAttribution } from './attributionProvider';
+import { getBvvAttribution } from './attribution.provider';
 
 /**
  * A function that returns a promise with an array of geoResources.
