+++ test/service/provider/attribution.provider.test.js
@@ -1,6 +1,6 @@
 import { $injector } from '../../../src/injection';
 import { AggregateGeoResource, GeoResource, GeoResourceTypes } from '../../../src/services/domain/geoResources';
-import { getBvvAttribution, getDefaultAttribution, getMinimalAttribution } from '../../../src/services/provider/attributionProvider';
+import { getBvvAttribution, getDefaultAttribution, getMinimalAttribution } from '../../../src/services/provider/attribution.provider';
 
 describe('Attribution provider', () => {
